// frontend/libs/console/legacy-oss/src/lib/dataSources/types.ts
export interface BaseTableColumn {
  column_name: string;
  data_type: string;
  data_type_name?: string;
}
// ...
export interface BaseTable {
  table_name: string;
  table_schema: string;
  columns: BaseTableColumn[];
  is_enum?: boolean;
}